Autostart from cli on Windows
